Create goldenyears.txt
Stories from the times after the last of the sensible founding generation departed (raz, caz, anton) but before the LiveJournal Generation took over.

It's probably fair to say that The Golden Years kept reminding everyone we were the largest non-barbecue club (i.e. had far less social focus, activities very online when no other clubs had a website let alone a chat room — unless we provided their website space) and yet, the biggest memories are that this is when things got the most social and the least technical.

That's not to underplay the hard work from all the technical volunteers keeping the servers up.  "We are not an ISP" but websites and accounts were into the second decade of awesome membership benefit.  It was a constant tension, as was "we are also not the Linux Server On As Many Types of Hardware as Possible Society", and yet …